requirements.txt
Pillow 5.0.0
